move constants into shared
