+++ backend/src/config/const.ts
@@ -1,3 +1,1 @@
 export const LOG4JS_BASE_CATEGORY_NAME = 'backend'
-export const FRONTEND_LOGIN_ROUTE = 'login'
-export const GRADIDO_REALM = 'gradido'
